Ensure YouTube videos only load on click (Fixes #14201)
