+++ bedrock/careers/templates/careers/diversity.html
@@ -312,5 +312,5 @@
 
 {% block js %}
   {{ js_bundle('protocol-modal') }}
-  {{ js_bundle('careers-hero-video') }}
+  {{ js_bundle('video-modal-embed') }}
 {% endblock %}
