+++ bedrock/careers/templates/careers/includes/big-picture-video.html
@@ -5,13 +5,15 @@
 #}
 
 <div class="c-hero-video">
-  <a class="video-play js-video-play" href="https://youtu.be/_obdx3VUzuQ" title="Play the video" role="button">
-    <img src="{{ static('img/careers/careers-video-thumbnail.jpg') }}" alt="" width="2430" height="1390">
-  </a>
-
-  <div class="mzp-u-modal-content hidden">
-    <div class="mzp-c-video">
-      <div class="video-player" data-video-id="_obdx3VUzuQ" data-video-title="Mozilla: The Big Picture"></div>
-    </div>
-  </div>
+  {{ video_modal_embed(
+    id='_obdx3VUzuQ',
+    title='Mozilla: The Big Picture',
+    image=resp_img(
+      url='img/careers/careers-video-thumbnail.jpg',
+      optional_attributes={
+        'width': '684',
+        'height': '385'
+      }
+    ),
+  ) }}
 </div>
