+++ bedrock/firefox/templates/firefox/features/picture-in-picture.html
@@ -4,6 +4,7 @@
  file, You can obtain one at https://mozilla.org/MPL/2.0/.
 #}
 
+{% from "macros.html" import video_inline_embed with context %}
 {% from "macros-protocol.html" import split, picto with context %}
 
 {% extends "firefox/features/base-article.html" %}
@@ -25,18 +26,26 @@
 <p>{{ ftl('features-pip-got-things-to-do-do-both-with') }}</p>
 
 <figure class="c-article-figure">
+  {# YouTube is blocked in China so zh-CN gets an alternative, self-hosted video. #}
+  {% if LANG == 'zh-CN' %}
   <div class="mzp-c-video">
-    {# YouTube is blocked in China so zh-CN gets an alternative, self-hosted video. #}
-    {% if LANG == 'zh-CN' %}
-      <video preload="metadata" controls poster="{{ static('img/firefox/features/pip/video-poster.jpg') }}" data-video-title="Red Panda Cubs - Firefox + Woodland Park Zoo">
-        <source src="https://assets.mozilla.net/video/red-pandas.webm" type="video/webm">
-      </video>
-    {% else %}
-      <a class="video-play js-video-play" href="https://youtu.be/F-nFQryDB0s" data-id="F-nFQryDB0s" data-video-title="Red Panda Cubs - Firefox + Woodland Park Zoo" title="{{ ftl('features-pip-play-the-video') }}">
-        <img src="{{ static('img/firefox/features/pip/video-poster.jpg') }}" alt="" width="1024" height="576">
-      </a>
-    {% endif %}
+    <video preload="metadata" controls poster="{{ static('img/firefox/features/pip/video-poster.jpg') }}" data-video-title="Red Panda Cubs - Firefox + Woodland Park Zoo">
+      <source src="https://assets.mozilla.net/video/red-pandas.webm" type="video/webm">
+    </video>
   </div>
+  {% else %}
+    {{ video_inline_embed(
+      id='F-nFQryDB0s',
+      title='Red Panda Cubs - Firefox + Woodland Park Zoo',
+      image=resp_img(
+        url='img/firefox/features/pip/video-poster.jpg',
+        optional_attributes={
+          'width': '592',
+          'height': '341'
+        }
+      ),
+    ) }}
+  {% endif %}
 </figure>
 
 <h3>{{ ftl('features-pip-heres-how-it-works') }}</h3>
@@ -90,5 +99,5 @@
 
 {% block js %}
   {{ super() }}
-  {{ js_bundle('firefox-features-pip') }}
+  {{ js_bundle('video-inline-embed') }}
 {% endblock %}
