+++ bedrock/firefox/templates/firefox/features/tips/tips.html
@@ -242,5 +242,6 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('firefox_tips') }}
+  {{ js_bundle('protocol-modal') }}
+  {{ js_bundle('video-modal-embed') }}
 {% endblock %}
