+++ bedrock/foundation/templates/foundation/annualreport/2019/index.html
@@ -1187,7 +1187,9 @@
   <section data-modal-parent="mozilla-builders-are-moving-fast-and-fixing-things" class="mzp-c-card-feature mzp-has-aspect-16-9 mzp-l-card-feature-left-half">
     <div class="mzp-c-card-feature-media-wrapper">
       <div class="mzp-c-card-feature-media">
-        <iframe width="560" height="315" src="https://www.youtube-nocookie.com/embed/OarBwHc2JzY?rel=0&showinfo=0&modestbranding=1" title="Watch the Mozilla Builders 2020 Final" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
+        <a class="video-play" href="https://youtu.be/OarBwHc2JzY" title="Watch the Mozilla Builders 2020 Final" target="_blank" rel="noopener noreferrer">
+          <img src="{{ static('img/foundation/annualreport/2019/mozilla-builders.jpg') }}" alt="" width="560" height="315">
+        </a>
       </div>
     </div>
     <div class="mzp-c-card-feature-content">
