+++ bedrock/mozorg/templates/mozorg/diversity/2021/index.html
@@ -4,11 +4,10 @@
  file, You can obtain one at https://mozilla.org/MPL/2.0/.
 #}
 
+{% from "macros.html" import sub_nav, video_inline_embed with context %}
+{% from "macros-protocol.html" import split, picto, card, with context %}
+
 {% extends "base-protocol-mozilla.html" %}
-
-{% from "macros.html" import sub_nav with context %}
-{% from "macros-protocol.html" import split, picto, card, with context %}]
-
 
 {% block page_title %}Diversity & Inclusion at Mozilla: 2021 Results{% endblock %}
 {% block page_desc %}Mozilla strives to ensure that who we are, what we build, and how we engage the world around us reflect our commitment to diversity, equity, and inclusion.{% endblock %}
@@ -84,10 +83,18 @@
       <p><strong>As an organization with employees, volunteers, and community members spanning the globe, we strive to ensure that who we are, what we build, and how we choose to engage the world around us reflect our commitment to diversity, equity, and inclusion — and that all feel empowered to join our movement.</strong></p>
       <p>Our annual disclosure gives us an opportunity to both reflect on our ambition to be a more diverse and inclusive organization and to hold ourselves accountable to the commitments we've made. In 2021, we leveraged the momentum from 2020 to build a more representative workforce and an inclusive culture internally and externally. We did this by uplifting voices that have been historically marginalized and by building a better internet that works for everyone. </p>
       <p>Leaning into our own culture of technical experimentation we piloted programs, captured feedback, and refined and scaled efforts, all in hopes of increasing our impact. With a spirit of curiosity and an openness to innovation, we ran over 100 experimental programs and initiatives of varying scale, complexity, and size. Our experiments included launching a 6-month program aimed at <a href="https://blog.mozilla.org/careers/mozilla-inclusion-champions/" target="_blank" rel="external noopener">building skills for allyship</a>, partnering with podcasts to <a href="https://getpocket.com/collections/behind-the-worlds-first-viral-video" target="_blank" rel="external noopener">amplify historic events</a>, expanding partnerships with HBCUs to explore ethical computing and apply an <a href="https://foundation.mozilla.org/en/blog/using-education-and-intersectionality-to-fight-bias/">intersectional lens to bias in AI</a>, and increasing investment in a venture capital fund supporting women and minority-led businesses. Each effort enabled us to learn more about how best to serve the needs of our organization and our communities. These actions enabled over 65% of our organization to meaningfully engage with DEI efforts and make progress in their own learning journey. </p>
-      <div class="mzp-c-video">
-        <iframe width="640" height="360" src="https://www.youtube-nocookie.com/embed/oDifVXtL10w" title="Watch the Diversity and Inclusion at Mozilla video" frameborder="0" allowfullscreen>
-        </iframe>
-      </div>
+
+      {{ video_inline_embed(
+        id='oDifVXtL10w',
+        title='Watch the Diversity and Inclusion at Mozilla video',
+        image=resp_img(
+          url='img/mozorg/diversity/2021/diversity-video-thumbnail.png',
+          optional_attributes={
+            'width': '928',
+            'height': '522'
+          }
+        ),
+      ) }}
     </div>
     <div class="c-section-wrapper">
       <h3><span class="highlight black">Looking back: progress across Mozilla</span></h3>
@@ -220,5 +227,7 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('protocol-modal') }}
+  {{ js_bundle('video-inline-embed') }}
   {{ js_bundle('diversity-2021') }}
 {% endblock %}
