+++ bedrock/mozorg/templates/mozorg/rise25/includes/hero-video.html
@@ -23,10 +23,10 @@
       }
     ) }}
   </a>
-</div>
 
-<div class="mzp-u-modal-content hidden">
-  <div class="mzp-c-video">
-    <div class="video-player" data-video-id="LbGtijF8qb0" data-video-title="Rise25 Awards: Call for Nominations"></div>
+  <div class="mzp-u-modal-content hidden">
+    <div class="mzp-c-video js-video-content">
+      <div class="video-player-frame" data-video-id="LbGtijF8qb0" data-video-title="Rise25 Awards: Call for Nominations"></div>
+    </div>
   </div>
 </div>
