+++ bedrock/mozorg/templates/mozorg/try-picture-in-picture.html
@@ -4,6 +4,7 @@
   file, You can obtain one at https://mozilla.org/MPL/2.0/.
 #}
 
+{% from "macros.html" import video_inline_embed with context %}
 {% from "macros-protocol.html" import picto with context %}
 
 {% extends "firefox/base/base-protocol.html" %}
@@ -13,10 +14,10 @@
 
 {% if LANG == 'en-US' %}
   {% set video_link = '2CT-zB6AO6k' %}
-  {% set video_thumbnail = static('img/mozorg/try-picture-in-picture/video-thumbnail.jpg') %}
+  {% set video_thumbnail = 'img/mozorg/try-picture-in-picture/video-thumbnail.jpg' %}
 {% else %}
   {% set video_link = 'F-nFQryDB0s' %}
-  {% set video_thumbnail = static('img/mozorg/try-picture-in-picture/video-thumbnail-alt.jpg') %}
+  {% set video_thumbnail = 'img/mozorg/try-picture-in-picture/video-thumbnail-alt.jpg' %}
 {% endif %}
 
 {% block page_css %}
@@ -28,11 +29,18 @@
 {% block content %}
 <section class="try-pip-content-main">
   <div class="mzp-l-content mzp-t-content-md mzp-t-content-nospace">
-    <div class="mzp-c-video">
-      <a class="video-play js-video-play" href="https://youtu.be/{{ video_link }}" data-id="{{ video_link }}" data-video-title=" Must-have add-ons you’re missing out on | #AskFirefox" title="Play the video">
-        <img src="{{ video_thumbnail }}" alt="" width="1024" height="576">
-      </a>
-    </div>
+    {{ video_inline_embed(
+      id=video_link,
+      title='Must-have add-ons you’re missing out on | #AskFirefox',
+      start='169' if LANG.startswith('en-') else '0',
+      image=resp_img(
+        url=video_thumbnail,
+        optional_attributes={
+          'width': '688',
+          'height': '387'
+        }
+      ),
+    ) }}
 
     <h2 class="try-pip-main-title">{{ ftl('try-pip-pop-out') }}</h2>
     <p class="try-pip-main-tagline has-pip-icon">{{ ftl('try-pip-pin-a-video') }}</p>
@@ -85,5 +93,5 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('try-picture-in-picture') }}
+  {{ js_bundle('video-inline-embed') }}
 {% endblock %}
