+++ media/css/careers/styles.scss
@@ -14,7 +14,6 @@
 @use './components/footer.scss';
 
 @use '~@mozilla-protocol/core/protocol/css/components/modal';
-@use '~@mozilla-protocol/core/protocol/css/components/video';
 @use '~@mozilla-protocol/core/protocol/css/components/callout';
 @use '~@mozilla-protocol/core/protocol/css/templates/card-layout';
 @use '../protocol/components/custom-menu-list';
