+++ media/css/firefox/facebook-container.scss
@@ -9,7 +9,7 @@ $brand-theme: 'firefox';
 @import '~@mozilla-protocol/core/protocol/css/includes/lib';
 @import '~@mozilla-protocol/core/protocol/css/components/logos/logo';
 @import '~@mozilla-protocol/core/protocol/css/components/logos/logo-product-firefox';
-@import '~@mozilla-protocol/core/protocol/css/components/video';
+@import '../protocol/components/video';
 
 // * -------------------------------------------------------------------------- */
 // Components
