+++ media/css/foundation/annual-report-2019-2020.scss
@@ -8,7 +8,7 @@ $image-path: '/media/protocol/img';
 @import '~@mozilla-protocol/core/protocol/css/includes/lib';
 @import '~@mozilla-protocol/core/protocol/css/components/feature-card';
 @import '~@mozilla-protocol/core/protocol/css/components/modal';
-@import '~@mozilla-protocol/core/protocol/css/components/video';
+@import '../protocol/components/video';
 
 // * -------------------------------------------------------------------------- */
 // Hero image & banner heading
@@ -372,6 +372,10 @@ $image-path: '/media/protocol/img';
                 color: var(--link-color-hover);
             }
         }
+    }
+
+    .video-play {
+        position: static;
     }
 
     .mzp-c-card-feature-content-container {
