+++ media/css/mozorg/diversity/diversity.scss
@@ -8,7 +8,7 @@ $image-path: '/media/protocol/img';
 @import '~@mozilla-protocol/core/protocol/css/includes/lib';
 @import '~@mozilla-protocol/core/protocol/css/components/feature-card';
 @import '~@mozilla-protocol/core/protocol/css/components/modal';
-@import '~@mozilla-protocol/core/protocol/css/components/video';
+@import '../../protocol/components/video';
 
 // * -------------------------------------------------------------------------- */
 // Hero image & banner heading
