+++ media/css/protocol/components/_video.scss
@@ -2,19 +2,14 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-$font-path: '/media/protocol/fonts';
-$image-path: '/media/protocol/img';
-
 @import '~@mozilla-protocol/core/protocol/css/includes/lib';
-@import '~@mozilla-protocol/core/protocol/css/components/modal';
-@import '~@mozilla-protocol/core/protocol/css/components/section-heading';
-@import '~@mozilla-protocol/core/protocol/css/components/zap';
-@import '~@mozilla-protocol/core/protocol/css/components/logos/logo';
-@import '~@mozilla-protocol/core/protocol/css/components/logos/logo-product-mozilla';
-@import '../../protocol/components/video';
+@import '~@mozilla-protocol/core/protocol/css/components/video';
 
-// Play button on videos
-.mzp-c-card-media-wrapper {
+.video-play {
+    display: block;
+    line-height: 0;
+    position: relative;
+
     &::after {
         background: url('/media/img/icons/video-play.svg') top left no-repeat;
         content: '';
@@ -29,9 +24,22 @@ $image-path: '/media/protocol/img';
         z-index: 2;
     }
 
-    &:hover::after,
-    &:focus::after {
-        opacity: 1;
-        transform: scale(1.1);
+    &:hover,
+    &:focus {
+        border-color: $color-white;
+        outline: 0;
+
+        &::after {
+            opacity: 1;
+            transform: scale(1.1);
+        }
+    }
+
+    .mzp-c-video & {
+        height: 100%;
+        left: 0;
+        position: absolute;
+        top: 0;
+        width: 100%;
     }
 }
