+++ media/js/firefox/whatsnew/whatsnew-120-na.es6.js
@@ -4,113 +4,10 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-/* eslint no-unused-vars: [2, { "varsIgnorePattern": "onYouTubeIframeAPIReady" }] */
-
 import MzpModal from '@mozilla-protocol/core/protocol/js/modal';
 
-const href = window.location.href;
-const videoLink = document.querySelector('.js-video-play');
-const src = 'https://www.youtube.com/iframe_api';
 const qrButton = document.querySelector('.qr-code-btn');
 const modalContent = document.querySelector('.mzp-u-modal-content');
-
-if (href.indexOf('v=1') !== -1) {
-    init();
-}
-
-// YouTube API hook has to be in global scope
-window.onYouTubeIframeAPIReady = () => {
-    // Play the video only once the API is ready.
-    Mozilla.pipVideoPlay();
-};
-
-function loadScript() {
-    const tag = document.createElement('script');
-    tag.src = src;
-    const firstScriptTag = document.getElementsByTagName('script')[0];
-    firstScriptTag.parentNode.insertBefore(tag, firstScriptTag);
-}
-
-function isScriptLoaded() {
-    return document.querySelector('script[src="' + src + '"]') ? true : false;
-}
-
-function playVideo() {
-    const videoId = videoLink.getAttribute('data-id');
-    const title = videoLink.getAttribute('data-video-title');
-
-    // return early if youtube API fails to load or is blocked.
-    if (typeof window.YT === 'undefined') {
-        return;
-    }
-
-    new window.YT.Player(videoLink, {
-        width: 500,
-        height: 281,
-        videoId: videoId,
-        playerVars: {
-            modestbranding: 1, // hide YouTube logo.
-            rel: 0, // do not show related videos on end.
-            cc_load_policy: 1 // show captions.
-        },
-        events: {
-            onReady: onPlayerReady,
-            onStateChange: onPlayerStateChange
-        }
-    });
-
-    function onPlayerReady(event) {
-        event.target.playVideo();
-    }
-
-    function onPlayerStateChange(event) {
-        let state;
-
-        switch (event.data) {
-            case window.YT.PlayerState.PLAYING:
-                state = 'video play';
-                break;
-            case window.YT.PlayerState.PAUSED:
-                state = 'video paused';
-                break;
-            case window.YT.PlayerState.ENDED:
-                state = 'video complete';
-                break;
-        }
-
-        if (state) {
-            window.dataLayer.push({
-                event: 'video-interaction',
-                videoTitle: title,
-                interaction: state
-            });
-        }
-    }
-}
-
-function initVideoPlayer() {
-    // check to see if you youtube API is loaded before trying to play the video.
-    if (!isScriptLoaded()) {
-        loadScript();
-    } else {
-        playVideo();
-    }
-}
-
-function init() {
-    videoLink.setAttribute('role', 'button');
-
-    videoLink.addEventListener(
-        'click',
-        function (e) {
-            e.preventDefault();
-            initVideoPlayer();
-        },
-        false
-    );
-}
-
-Mozilla.pipVideoPlay = playVideo;
 
 // click handler for opening Modal with QR Code
