+++ media/js/mozorg/rise25/rise25.js
@@ -4,16 +4,6 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-/* eslint no-unused-vars: [2, { "varsIgnorePattern": "onYouTubeIframeAPIReady" }] */
-
-// YouTube API hook has to be in global scope, ugh.
-window.onYouTubeIframeAPIReady = function () {
-    'use strict';
-
-    // Play the video only once the API is ready.
-    Mozilla.heroVideoPlay();
-};
-
 (function () {
     'use strict';
 
@@ -106,123 +96,5 @@ window.onYouTubeIframeAPIReady = function () {
     for (var index = 0; index < copyLinks.length; index++) {
         var element = copyLinks[index];
         element.addEventListener('click', handleCopyLink);
-    }
-
-    // Video
-    var heroVideo = document.querySelector('.r25-hero .video-player');
-    var videoTitle = heroVideo.getAttribute('data-video-title');
-    var videoId = heroVideo.getAttribute('data-video-id');
-    var src = 'https://www.youtube.com/iframe_api';
-    var player;
-
-    function loadScript() {
-        var tag = document.createElement('script');
-        tag.src = src;
-        var firstScriptTag = document.getElementsByTagName('script')[0];
-        firstScriptTag.parentNode.insertBefore(tag, firstScriptTag);
-    }
-
-    function isScriptLoaded() {
-        return document.querySelector('script[src="' + src + '"]')
-            ? true
-            : false;
-    }
-
-    function playVideo() {
-        // if youtube API fails or is blocked, try redirecting to youtube.com
-        if (typeof window.YT === 'undefined') {
-            window.location.href = 'https://www.youtube.com/watch?v=' + videoId;
-        }
-
-        player = new window.YT.Player(heroVideo, {
-            width: 640,
-            height: 360,
-            videoId: videoId,
-            playerVars: {
-                modestbranding: 1, // hide YouTube logo.
-                rel: 0 // do not show related videos on end.
-            },
-            events: {
-                onReady: onPlayerReady,
-                onStateChange: onPlayerStateChange
-            }
-        });
-
-        function onPlayerReady(event) {
-            event.target.playVideo();
-        }
-
-        function onPlayerStateChange(event) {
-            var state;
-
-            switch (event.data) {
-                case window.YT.PlayerState.PLAYING:
-                    state = 'video play';
-                    break;
-                case window.YT.PlayerState.PAUSED:
-                    state = 'video paused';
-                    break;
-                case window.YT.PlayerState.ENDED:
-                    state = 'video complete';
-                    break;
-            }
-
-            if (state) {
-                window.dataLayer.push({
-                    event: 'video-interaction',
-                    videoTitle: videoTitle,
-                    interaction: state
-                });
-            }
-        }
-    }
-
-    function initVideoPlayer() {
-        // check to see if you youtube API is loaded before trying to play the video.
-        if (!isScriptLoaded()) {
-            loadScript();
-        } else {
-            playVideo();
-        }
-    }
-
-    function destroyVideoPlayer() {
-        if (player) {
-            player.destroy();
-        }
-    }
-
-    function openVideoModal(e) {
-        e.preventDefault();
-
-        var content = document.querySelector('.mzp-u-modal-content');
-
-        window.MzpModal.createModal(e.target, content, {
-            title: videoTitle,
-            className: 'mzp-has-media',
-            onCreate: initVideoPlayer,
-            onDestroy: destroyVideoPlayer
-        });
-    }
-
-    function initVideo() {
-        var heroVideoButtons = document.querySelectorAll(
-            '.r25-hero .js-video-play'
-        );
-
-        for (var i = 0; i < heroVideoButtons.length; i++) {
-            heroVideoButtons[i].setAttribute('role', 'button');
-            heroVideoButtons[i].addEventListener(
-                'click',
-                openVideoModal,
-                false
-            );
-        }
-    }
-
-    Mozilla.heroVideoPlay = playVideo;
-
-    if (heroVideo) {
-        initVideo();
     }
 })();
