+++ media/static-bundles.json
@@ -1445,12 +1445,6 @@
     },
     {
       "files": [
-        "js/firefox/features/tips/tips-youtube.js"
-      ],
-      "name": "firefox_tips"
-    },
-    {
-      "files": [
         "js/firefox/features/features-article.es6.js"
       ],
       "name": "firefox_features_article"
@@ -1670,12 +1664,6 @@
         "js/firefox/privacy/products.js"
       ],
       "name": "firefox-privacy-products"
-    },
-    {
-      "files": [
-        "js/firefox/features/picture-in-picture.js"
-      ],
-      "name": "firefox-features-pip"
     },
     {
       "files": [
@@ -1899,9 +1887,15 @@
     },
     {
       "files": [
-        "js/careers/hero-video.js"
+        "js/base/video-inline-embed.es6.js"
       ],
-      "name": "careers-hero-video"
+      "name": "video-inline-embed"
+    },
+    {
+      "files": [
+        "js/base/video-modal-embed.es6.js"
+      ],
+      "name": "video-modal-embed"
     },
     {
       "files": [
@@ -1964,12 +1958,6 @@
         "js/products/mozsocial/invite.es6.js"
       ],
       "name": "mozsocial-waitlist"
-    },
-    {
-      "files": [
-        "js/mozorg/try-picture-in-picture.js"
-      ],
-      "name": "try-picture-in-picture"
     },
     {
       "files": [
